issue 19

+++ src/controller/shipping.controller.ts
@@ -16,9 +16,13 @@ export class ShippingController {
     return 'pong';
   }
 
-  @Post('shipping/:orderId')
+  @Post(':orderId')
   async shipOrder(@Param('orderId') orderId: string) {
     const orderProducts = await this.stockService.getOrderProducts(orderId);
+    if (orderProducts.length === 0) {
+      throw new BadRequestException(`La commande ${orderId} n'existe pas.`);
+    }
+
     for (const product of orderProducts) {
       const success = await this.stockService.removeStock(
         product.productId,
@@ -26,10 +30,10 @@ export class ShippingController {
       );
       if (!success) {
         throw new BadRequestException(
-          `Stock insuffisant pour le produit ${product.productId}`,
+          `Stock insuffisant pour le produit ${product.productId}.`,
         );
       }
     }
-    return { message: `Commande ${orderId} expédiée` };
+    return { message: `Produits de la commande ${orderId} supprimé des stock` };
   }
 }
